Added short format output
